+++ Runtime/SpriteAnimationState.cs
@@ -1,8 +1,0 @@
-﻿namespace GabrielBigardi.SpriteAnimator.Runtime
-{
-    public enum SpriteAnimationState
-    {
-        Playing = 0,
-        Paused = 1
-    }
-}