+++ Runtime/SpriteAnimationType.cs
@@ -1,8 +1,0 @@
-﻿namespace GabrielBigardi.SpriteAnimator.Runtime
-{
-    public enum SpriteAnimationType
-    {
-        Looping = 0,
-        PlayOnce = 1
-    }
-}