+++ Runtime/SpriteAnimator.cs
@@ -1,226 +1,126 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
+using System;
 using UnityEngine;
-using Debug = UnityEngine.Debug;
 
-namespace GabrielBigardi.SpriteAnimator.Runtime
+namespace GabrielBigardi.SpriteAnimator
 {
     public class SpriteAnimator : MonoBehaviour
     {
-        [SerializeField] private List<SpriteAnimation> _spriteAnimations = new();
         [SerializeField] private bool _playAutomatically = true;
-        [SerializeField] private bool _debugMode = false;
+        [SerializeField] private SpriteAnimationObject _spriteAnimationObject;
+        [SerializeField] private SpriteRenderer _spriteRenderer;
 
-        public SpriteAnimation DefaultAnimation => _spriteAnimations.Count > 0 ? _spriteAnimations[0] : null;
-        public SpriteAnimation CurrentAnimation => SpriteAnimationHelper.CurrentAnimation;
+        private SpriteAnimation _currentAnimation;
+        private float _animationTime = 0.0f;
+        private bool _paused = false;
 
-        public bool Playing => _state == SpriteAnimationState.Playing;
-        public bool Paused => _state == SpriteAnimationState.Paused;
-        public int CurrentFrame => SpriteAnimationHelper.GetCurrentFrame();
-        public bool IsLastFrame => CurrentFrame == CurrentAnimation.Frames.Count - 1;
+        public bool HasAnimation(string name) => _spriteAnimationObject.SpriteAnimations.Exists(a => a.Name == name);
+        public SpriteAnimation DefaultAnimation => _spriteAnimationObject.SpriteAnimations.Count > 0 ? _spriteAnimationObject.SpriteAnimations[0] : null;
+        private int GetLoopingFrame() => (int)_animationTime % _currentAnimation.Frames.Count;
+        private int GetPlayOnceFrame() => Mathf.Min((int)_animationTime, _currentAnimation.Frames.Count - 1);
 
-        private SpriteRenderer _spriteRenderer;
-        public SpriteAnimationHelper SpriteAnimationHelper { get; private set; }
-        private SpriteAnimationState _state = SpriteAnimationState.Playing;
-        private SpriteAnimationFrame _previousAnimationFrame;
+        // Events
+        public event Action OnAnimationComplete;
 
-        private bool triggerAnimationEndedEvent = false;
-
-        public event Action SpriteChanged;
-        public event Action<SpriteAnimation> AnimationPlayed;
-        public event Action<SpriteAnimation> AnimationPaused;
-        public event Action<SpriteAnimation> AnimationEnded;
-        public event Action<string> AnimationEventCalled;
-
-        private void Awake()
-        {
-            SpriteAnimationHelper = new SpriteAnimationHelper();
-
-            _spriteRenderer = GetComponentInChildren<SpriteRenderer>();
-            if (_spriteRenderer == null)
-            {
-                Debug.LogError("No Sprite Renderers were found on the object or the child objects");
-                return;
-            }
-        }
-
-        private void Start()
+        private void OnEnable()
         {
             if (_playAutomatically)
             {
-                if(DefaultAnimation == null)
+                if (_spriteAnimationObject.SpriteAnimations[0] == null)
                 {
-                    Debug.LogError($"PlayAutomatically is True but no default animations were found, drag a animation into SpriteAnimations list");
+                    Debug.LogError($"PlayAutomatically is set to TRUE but no default animations were found.");
                     return;
                 }
 
-                DebugModeMessage($"Playing default animation {DefaultAnimation.name}");
                 Play(DefaultAnimation);
             }
         }
 
         private void LateUpdate()
         {
-            if (!Playing || _spriteRenderer == null)
+            if (_paused || _spriteRenderer == null)
                 return;
 
-            SpriteAnimationFrame currentFrame = SpriteAnimationHelper.UpdateAnimation(Time.deltaTime);
+            Sprite currentFrame = UpdateAnimation(Time.deltaTime);
 
             if (currentFrame == null)
                 return;
 
-            if (currentFrame == _previousAnimationFrame)
-                return;
-
-            if (triggerAnimationEndedEvent)
-            {
-                DebugModeMessage($"Calling AnimationEnded event");
-                AnimationEnded?.Invoke(CurrentAnimation);
-                triggerAnimationEndedEvent = false;
-
-                if (CurrentAnimation.SpriteAnimationType != SpriteAnimationType.Looping)
-                {
-                    DebugModeMessage($"Reached end of non-loopable animation {CurrentAnimation.name}");
-                    return;
-                }
-            }
-
-            if ((CurrentFrame + 1) > (CurrentAnimation.Frames.Count - 1))
-            {
-                triggerAnimationEndedEvent = true;
-            }
-
-            _previousAnimationFrame = currentFrame;
-
-            _spriteRenderer.sprite = currentFrame.Sprite;
-
-            SpriteChanged?.Invoke();
-            if (currentFrame.EventName != "")
-            {
-                AnimationEventCalled?.Invoke(currentFrame.EventName);
-                DebugModeMessage($"Calling animation frame event {currentFrame.EventName}");
-            }
+            _spriteRenderer.sprite = currentFrame;
         }
 
-        public bool HasAnimation(string name)
-        {
-            return _spriteAnimations.Exists(a => a.Name == name);
-        }
-
-        public void PlayIfNotPlaying(string name)
+        public SpriteAnimator Play(string name, int startFrame = 0)
         {
             if (!HasAnimation(name))
             {
                 Debug.LogError($"Animation with name '{name}' not found");
-                return;
+                return null;
             }
 
-            PlayIfNotPlaying(GetAnimationByName(name));
-        }
-
-        public void PlayIfNotPlaying(SpriteAnimation animation)
-        {
-            if (CurrentAnimation.name != animation.name)
-            {
-                _state = SpriteAnimationState.Playing;
-                SpriteAnimationHelper.ChangeAnimation(animation);
-                AnimationPlayed?.Invoke(animation);
-
-                triggerAnimationEndedEvent = false;
-
-                DebugModeMessage($"Playing animation {animation.name}");
-            }
-        }
-
-        public void Play()
-        {
-            if (CurrentAnimation == null)
-            {
-                SpriteAnimationHelper.ChangeAnimation(DefaultAnimation);
-            }
-
-            Play(CurrentAnimation);
-        }
-
-        public void Play(string name)
-        {
-            if (!HasAnimation(name))
-            {
-                Debug.LogError($"Animation with name '{name}' not found");
-                return;
-            }
-
-            Play(GetAnimationByName(name));
-        }
-
-        public void Play(SpriteAnimation animation)
-        {
-            if (animation == null)
-            {
-                Debug.LogError("An null or invalid SpriteAnimation object was passed.");
-                return;
-            }
-
-            _state = SpriteAnimationState.Playing;
-            SpriteAnimationHelper.ChangeAnimation(animation);
-            AnimationPlayed?.Invoke(animation);
-
-            triggerAnimationEndedEvent = false;
-
-            DebugModeMessage($"Playing animation {animation.name}");
-        }
-
-        public void Play(SpriteAnimation animation, int startFrame = 0)
-        {
-            if (animation == null)
-            {
-                Debug.LogError("An null or invalid SpriteAnimation object was passed.");
-                return;
-            }
-
-            if (CurrentAnimation == null || CurrentAnimation.name != animation.name)
-            {
-                _state = SpriteAnimationState.Playing;
-                SpriteAnimationHelper.ChangeAnimation(animation);
-                AnimationPlayed?.Invoke(animation);
-
-                triggerAnimationEndedEvent = false;
-            }
-
-            _state = SpriteAnimationState.Playing;
-            SpriteAnimationHelper.SetCurrentFrame(startFrame);
-
-            DebugModeMessage($"Playing animation {animation.name} starting on frame {startFrame}");
-        }
-
-        public void Play(string name, int startFrame = 0)
-        {
             Play(GetAnimationByName(name), startFrame);
+            return this;
         }
 
-        public void Pause()
+        public SpriteAnimator Play(SpriteAnimation spriteAnimation, int startFrame = 0)
         {
-            DebugModeMessage($"Pausing animator");
+            if (spriteAnimation == null)
+            {
+                Debug.LogError("An null or invalid SpriteAnimation object was passed.");
+                return null;
+            }
 
-            _state = SpriteAnimationState.Paused;
-            AnimationPaused?.Invoke(CurrentAnimation);
+            Resume();
+            ChangeAnimation(spriteAnimation);
+            SetCurrentFrame(startFrame);
+
+            return this;
         }
 
-        public void Resume()
+        public SpriteAnimator PlayIfNotPlaying(string name, int startFrame = 0)
         {
-            DebugModeMessage($"Resuming animator");
+            if (!HasAnimation(name))
+            {
+                Debug.LogError($"Animation with name '{name}' not found");
+                return null;
+            }
 
-            _state = SpriteAnimationState.Playing;
+            PlayIfNotPlaying(GetAnimationByName(name), startFrame);
+            return this;
+        }
+
+        public SpriteAnimator PlayIfNotPlaying(SpriteAnimation spriteAnimation, int startFrame = 0)
+        {
+            if (spriteAnimation == null)
+            {
+                Debug.LogError("An null or invalid SpriteAnimation object was passed.");
+                return null;
+            }
+
+            if (_currentAnimation.Name == spriteAnimation.Name)
+                return null;
+            Resume();
+            ChangeAnimation(spriteAnimation);
+            SetCurrentFrame(startFrame);
+
+            return this;
+        }
+
+        public void Pause() => _paused = true;
+
+        public void Resume() => _paused = false;
+
+        public SpriteAnimator OnComplete(Action onAnimationComplete)
+        {
+            this.OnAnimationComplete = onAnimationComplete;
+            return this;
         }
 
         public SpriteAnimation GetAnimationByName(string name)
         {
-            for (int i = 0; i < _spriteAnimations.Count; i++)
+            for (int i = 0; i < _spriteAnimationObject.SpriteAnimations.Count; i++)
             {
-                if (_spriteAnimations[i].Name == name)
+                if (_spriteAnimationObject.SpriteAnimations[i].Name == name)
                 {
-                    return _spriteAnimations[i];
+                    return _spriteAnimationObject.SpriteAnimations[i];
                 }
             }
 
@@ -228,12 +128,76 @@ namespace GabrielBigardi.SpriteAnimator.Runtime
             return null;
         }
 
-        private void DebugModeMessage(string message)
+        private void ChangeAnimation(SpriteAnimation spriteAnimation)
         {
-            if (!_debugMode)
-                return;
+            _animationTime = 0f;
+            OnAnimationComplete = null;
+            _currentAnimation = spriteAnimation;
+        }
 
-            Debug.Log(message);
+        public Sprite UpdateAnimation(float deltaTime)
+        {
+            if (_currentAnimation != null)
+            {
+                _animationTime += deltaTime * _currentAnimation.FPS;
+
+                var frameDuration = 1f / _currentAnimation.FPS;
+                var animationDuration = (frameDuration * (_currentAnimation.Frames.Count)) * 10;
+                if (_animationTime >= animationDuration)
+                {
+                    OnAnimationComplete?.Invoke();
+                    OnAnimationComplete = null;
+                }
+
+                return GetAnimationFrame();
+            }
+
+            return null;
+        }
+
+        private Sprite GetAnimationFrame()
+        {
+            int currentFrame = 0;
+
+            switch (_currentAnimation.SpriteAnimationType)
+            {
+                case SpriteAnimationType.Looping:
+                    currentFrame = GetLoopingFrame();
+                    break;
+                case SpriteAnimationType.PlayOnce:
+                    currentFrame = GetPlayOnceFrame();
+                    break;
+            }
+
+            return _currentAnimation.Frames[currentFrame];
+        }
+
+        public int GetCurrentFrame()
+        {
+            int currentFrame = 0;
+
+            switch (_currentAnimation.SpriteAnimationType)
+            {
+                case SpriteAnimationType.Looping:
+                    currentFrame = GetLoopingFrame();
+                    break;
+                case SpriteAnimationType.PlayOnce:
+                    currentFrame = GetPlayOnceFrame();
+                    break;
+            }
+
+            return currentFrame;
+        }
+
+        public void SetCurrentFrame(int frame)
+        {
+            if (frame < 0 || frame >= _currentAnimation.Frames.Count)
+            {
+                Debug.LogError($"Invalid frame index {frame} for animation '{_currentAnimation.Name}' with {_currentAnimation.Frames.Count} frames");
+                return;
+            }
+
+            _animationTime = frame;
         }
     }
 }